comandos/requirements.txt
pyjwt==2.0.0
datetime
uuid
Flask-SQLAlchemy
Flask-JWT-Extended==4.2.3